Fix #4357: Move Look Assigner tool to maya since it's Maya only
