+++ openpype/hosts/maya/tools/mayalookassigner/LICENSE
